+++ .github/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+-   Changed the attribute used to check already exisiting emails when updating one's account
+
 ## [2.0.0](https://github.com/hackmcgill/hackerapi/tree/2.0.0) - 2019-12-17
 
 ### Added
@@ -16,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Add accept hacker route
 -   Create changelog
 
-
 ### Changed
 
 -   Match npm scripts with dashboard
@@ -27,7 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Update pre-acceptance email templates: none hacker status, applied hacker status, account invitation, account confirmation, password reset
 -   Group application enums
 -   Change job interest enums
--   Changed the attribute used to check already exisiting emails when updating one's account
 
 ### Fixed