+++ .github/CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   Add tests for `travel` field in hacker's application
 -   Add email blast marketing email
+
+### Changed
+
+-   Removed all instances of `needsBus`
 
 ### Fixed