+++ .github/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Add tests for `travel` field in hacker's application
+-   Add email blast marketing email
 
 ### Changed