+++ .github/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1](https://github.com/hackmcgill/hackerapi/tree/2.0.1) - 2019-12-24
+
+### Added
+
+-   Add tests for `travel` field in hacker's application
+
+### Changed
+
+-   Removed all instances of `needsBus`
+
+### Fixed
+
+-   Change the attribute used to check already exisiting emails when updating one's account
+
 ## [2.0.0](https://github.com/hackmcgill/hackerapi/tree/2.0.0) - 2019-12-17
 
 ### Added